feat: initial extraction of git.rs
